Finished wrapping everything but Docking functions and Drag n drop functions.
Essentially this is done. It still needs to be heavily tested. Next step is to port the entire imgui_demo.cpp content to GML. If I need to wrap the remaining functions then I will. At the moment it seems like I won't need them. But I'm a big fat noob, so what do I know?