Finished up Trees, Collapsing Headers and Bullets sections.
Started working on Text section. But have to wrap ImDrawList functions for custom rendering and I just don't feel like doing that right now.

Drag and drop might actually be working fine. Apparently the tables aren't dropped anywhere in the original demo as well, so working as intended at the moment? We'll see once we get to use the payload data.